feat(CreatePostRequest): change price type from Double to BigDecimal for better precision in monetary values feat(Post): update price field type from Double to BigDecimal to ensure accurate representation of prices

+++ GE/src/main/java/org/example/ge/domain/post/controller/dto/request/CreatePostRequest.java
@@ -2,10 +2,12 @@ package org.example.ge.domain.post.controller.dto.request;
 
 import org.example.ge.domain.post.entity.Post;
 
+import java.math.BigDecimal;
+
 public record CreatePostRequest(
         String title,
         String content,
-        Double price
+        BigDecimal price
 ) {
     public Post toEntity(Long userId, String imageUrl) {
         return Post.builder()
